약간의 수정..

+++ Homeworks/HW6/hw6-1-1.c
@@ -14,20 +14,20 @@
 
 #include <stdio.h>
 
-int fact(int n);
+unsigned long long fact(unsigned long long n);
 
 int main(void) {
-    int n, trees;
-    scanf("%d", &n);
+    unsigned long long n, trees;
+    scanf("%llu", &n);
 
     // ₂nCn = 2n! / (2n-n)! n!
     trees = fact(2*n) / fact(2*n - n) * fact(n);
-    printf("trees: %d\n", trees);
+    printf("trees: %llu\n", trees);
 }
 
-int fact(int n) {
-    int result = 1;
-    for (int i = 1; i <= n; i++)
+unsigned long long fact(unsigned long long n) {
+    unsigned long long result = 1;
+    for (unsigned long long i = 1; i <= n; i++)
         result *= i;
     return result;
 }
